feat(postman) Add collection/environment for public api
This collection contains a full chain from oauth to call of each
of the API endpoints.

Perhaps this is better to have in some other place. But would be nice
to have publicly available, wouldn't it? I think it's really helpful
for us to provide postman, instead of users of our API needing to use
curl for exploring our API.

Any suggestions for better place to put this? Or should we avoid having
postman collections at all?